+++ languages/languageDesign/textGen/source_gen/jetbrains/mps/lang/textGen/typesystem/check_CallInnerFunctionOperation_NonTypesystemRule.java
@@ -28,7 +28,7 @@ public class check_CallInnerFunctionOperation_NonTypesystemRule extends Abstract
     if (SNodeOperations.isInstanceOf(SLinkOperations.getTarget(ListSequence.fromList(parameterDeclarations).last(), "type", true), "jetbrains.mps.baseLanguage.structure.VariableArityType")) {
       b = ListSequence.fromList(parameterDeclarations).count() - 1 <= ListSequence.fromList(actualArguments).count();
     } else {
-      b = (int) ListSequence.fromList(parameterDeclarations).count() == (int) ListSequence.fromList(actualArguments).count();
+      b = ListSequence.fromList(parameterDeclarations).count() == ListSequence.fromList(actualArguments).count();
     }
     if (!(b)) {
       {
