+++ plugins/vcs/common/source_gen/jetbrains/mps/vcs/platform/mergedriver/GitGlobalInstaller.java
@@ -95,7 +95,7 @@ import org.apache.log4j.LogManager;
           return !(line.trim().isEmpty());
         }
       }).toListSequence();
-      boolean equal = (int) ListSequence.fromList(section).count() == (int) ListSequence.fromList(newConfigLines).count();
+      boolean equal = ListSequence.fromList(section).count() == ListSequence.fromList(newConfigLines).count();
       if (equal) {
         for (int i = 0; i < ListSequence.fromList(section).count(); i++) {
           if (neq_btx4zt_a0a0a0g0t0d(ListSequence.fromList(section).getElement(i), ListSequence.fromList(newConfigLines).getElement(i))) {
